Move errors mod one up from control_plane::client
